Dispatches for enum creation/deletion and alteration joined into one to avoid multiple postgresql calls - Enum creation and deletion split into two files
