+++ alembic_postgresql_enum/__init__.py
@@ -1,2 +1,1 @@
-from .enum_creation import *
-from .enum_alteration import *
+from .compare_dispatch import *
