+++ alembic_postgresql_enum/enum_alteration.py
@@ -13,8 +13,7 @@ import sqlalchemy
 from alembic.autogenerate.api import AutogenContext
 from alembic.operations.ops import UpgradeOps
 
-from . import EnumNamesToValues
-from .get_enum_data import get_connection, get_defined_enums, get_declared_enums, EnumNamesToTableReferences
+from .get_enum_data import get_connection, EnumNamesToValues, EnumNamesToTableReferences
 
 
 @alembic.operations.base.Operations.register_operation("sync_enum_values")
@@ -124,24 +123,3 @@ def sync_changed_enums(defined_enums: EnumNamesToValues, declared_enums: EnumNam
         op = SyncEnumValuesOp(schema, enum_name, list(old_values), list(new_values),
                               [column_reference.to_tuple() for column_reference in affected_columns])
         upgrade_ops.ops.append(op)
-
-
-@alembic.autogenerate.comparators.dispatch_for("schema")
-def compare_enums(autogen_context: AutogenContext, upgrade_ops, schema_names):
-    """
-    Walk the declared SQLAlchemy schema for every referenced Enum, walk the PG
-    schema for every defined Enum, then generate SyncEnumValuesOp migrations
-    for each defined enum that has grown new entries when compared to its
-    declared version.
-    Enums that don't exist in the database yet are ignored, since
-    SQLAlchemy/Alembic will create them as part of the usual migration process.
-    """
-    for schema in schema_names:
-        default_schema = autogen_context.dialect.default_schema_name
-        if schema is None:
-            schema = default_schema
-
-        defined = get_defined_enums(autogen_context.connection, schema)
-        declared = get_declared_enums(autogen_context.metadata, schema, default_schema, autogen_context.dialect)
-
-        sync_changed_enums(defined, declared.enum_values, declared.enum_table_references, schema, upgrade_ops)
